requirements.txt
altair==5.5.0
appnope==0.1.3
asttokens==2.4.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.1.4
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.5.0
jupyter_core==5.5.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.6
narwhals==1.35.0
nest-asyncio==1.5.8
numpy==2.2.4
opencv-python==4.11.0.86
packaging==23.2
pandas==2.2.3
parso==0.8.3
pexpect==4.8.0
pillow==11.2.1
platformdirs==3.11.0
plotly==6.0.1
prompt-toolkit==3.0.39
protobuf==5.29.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.16.1
pyparsing==3.2.3
python-dateutil==2.8.2
pytz==2025.2
pyzmq==25.1.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scipy==1.15.2
six==1.16.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.9
